refactor: 여러군데 혼재 되어있던 유저관련 책임을 User모듈에게 위임
