+++ main-presentation/build.gradle.kts
@@ -25,6 +25,8 @@ dependencies {
     implementation(projects.settingPresentation)
     implementation(projects.settingDomain)
 
+    implementation(projects.userDomain)
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.androidx.activity.compose)
