+++ onboarding-data/src/main/java/com/teamhy2/onboarding/data/di/OnboardingRepositoryModule.kt
@@ -2,9 +2,7 @@ package com.teamhy2.onboarding.data.di
 
 import com.teamhy2.onboarding.data.repository.DefaultWebViewRepository
 import com.teamhy2.onboarding.data.repository.InMemoryDepartmentRepository
-import com.teamhy2.onboarding.data.repository.RemoteUserRepository
 import com.teamhy2.onboarding.domain.repository.DepartmentRepository
-import com.teamhy2.onboarding.domain.repository.UserRepository
 import com.teamhy2.onboarding.domain.repository.WebViewRepository
 import dagger.Binds
 import dagger.Module
@@ -18,10 +16,6 @@ abstract class OnboardingRepositoryModule {
     @Binds
     @Singleton
     abstract fun bindWebViewRepository(defaultWebViewRepository: DefaultWebViewRepository): WebViewRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindsUserRepository(remoteUserRepository: RemoteUserRepository): UserRepository
 
     @Binds
     @Singleton
