--- onboarding-domain/src/main/java/com/teamhy2/onboarding/domain/repository/UserRepository.kt
@@ -1,19 +1,0 @@
-package com.teamhy2.onboarding.domain.repository
-
-typealias AlreadyExist = Boolean
-typealias Duplication = Boolean
-
-interface UserRepository {
-    suspend fun checkNicknameDuplication(nickname: String): Result<Duplication>
-
-    suspend fun signUp(
-        nickname: String,
-        department: String,
-    ): Result<Unit>
-
-    suspend fun signIn(idToken: String): Result<AlreadyExist>
-
-    suspend fun signOut(): Result<Unit>
-
-    suspend fun withdraw(): Result<Unit>
-}
