--- setting-domain/src/main/java/com/teamhy2/feature/setting/domain/repository/model/UserInfo.kt
@@ -1,6 +1,0 @@
-package com.teamhy2.feature.setting.domain.repository.model
-
-data class UserInfo(
-    val name: String,
-    val department: String,
-)
