+++ user-domain/src/main/java/com/teamhy2/user/domain/repository/UserRepository.kt
@@ -2,6 +2,22 @@ package com.teamhy2.user.domain.repository
 
 import com.teamhy2.user.domain.model.UserInfo
 
+typealias AlreadyExist = Boolean
+typealias Duplication = Boolean
+
 interface UserRepository {
+    suspend fun checkNicknameDuplication(nickname: String): Result<Duplication>
+
+    suspend fun signUp(
+        nickname: String,
+        department: String,
+    ): Result<Unit>
+
+    suspend fun signIn(idToken: String): Result<AlreadyExist>
+
+    suspend fun signOut(): Result<Unit>
+
+    suspend fun withdraw(): Result<Unit>
+
     suspend fun getUserInfo(): Result<UserInfo>
 }
